requirements.txt
openai==1.8.0
PyYAML==5.4.1
tqdm==4.66.1
